requirements.txt
streamlit
audio-recorder-streamlit
SpeechRecognition
google-generativeai
gtts
python-dotenv
pyaudio